+++ generatorcore/agri2030/energy_demand/co2eChangeOther.py
@@ -1,0 +1,52 @@
+# pyright: strict
+
+from dataclasses import dataclass, InitVar
+
+from ...inputs import Inputs
+from ...agri2018.a18 import A18
+
+from .co2eChange import CO2eChange
+
+
+@dataclass(kw_only=True)
+class CO2eChangeOther(CO2eChange):
+    CO2e_production_based_per_t: float = 0
+    demand_change: float = 0
+    prod_volume: float = 0
+
+    inputs: InitVar[Inputs]
+    what: InitVar[str]
+    a18: InitVar[A18]
+    ass_demand_change: InitVar[str]
+    fact_production_based_per_t: InitVar[str]
+
+    def __post_init__(  # type: ignore
+        self,
+        inputs: Inputs,
+        what: str,
+        a18: A18,
+        ass_demand_change: str,
+        fact_production_based_per_t: str,
+    ):
+
+        self.CO2e_combustion_based = 0
+
+        self.demand_change = inputs.ass(ass_demand_change)
+        self.prod_volume = getattr(a18, what).prod_volume * (1 + self.demand_change)
+
+        self.CO2e_production_based_per_t = inputs.fact(fact_production_based_per_t)
+        self.CO2e_production_based = self.prod_volume * self.CO2e_production_based_per_t
+
+        parent = CO2eChange(
+            inputs=inputs,
+            what=what,
+            a18=a18,
+            CO2e_combustion_based=self.CO2e_combustion_based,
+            CO2e_production_based=self.CO2e_production_based,
+        )
+
+        self.CO2e_total = parent.CO2e_total
+        self.CO2e_total_2021_estimated = parent.CO2e_total_2021_estimated
+        self.change_CO2e_pct = parent.change_CO2e_pct
+        self.change_CO2e_t = parent.change_CO2e_t
+        self.cost_climate_saved = parent.cost_climate_saved
