+++ generatorcore/agri2030/energy_demand/co2eChangeSoil.py
@@ -1,0 +1,106 @@
+# pyright: strict
+
+from dataclasses import dataclass
+
+from ...inputs import Inputs
+from ...utils import div
+from ...agri2018.a18 import A18
+
+from .co2eChange import CO2eChange
+
+
+@dataclass(kw_only=True)
+class CO2eChangeSoil(CO2eChange):
+    CO2e_production_based_per_t: float = 0
+    area_ha: float = 0
+    area_ha_change: float = 0
+    demand_change: float = 0
+
+    @classmethod
+    def calc_soil_special(
+        cls,
+        inputs: Inputs,
+        what: str,
+        a18: A18,
+        area_ha: float,
+        CO2e_production_based_per_t: float,
+    ) -> "CO2eChangeSoil":
+
+        CO2e_combustion_based = 0
+
+        demand_change = (
+            div(
+                CO2e_production_based_per_t,
+                getattr(a18, what).CO2e_production_based_per_t,
+            )
+            - 1
+        )
+
+        CO2e_production_based = area_ha * CO2e_production_based_per_t
+
+        area_ha_change = -(getattr(a18, what).area_ha - area_ha)
+
+        parent = CO2eChange(
+            inputs=inputs,
+            what=what,
+            a18=a18,
+            CO2e_combustion_based=CO2e_combustion_based,
+            CO2e_production_based=CO2e_production_based,
+        )
+
+        return cls(
+            inputs=inputs,
+            what=what,
+            a18=a18,
+            CO2e_combustion_based=CO2e_combustion_based,
+            CO2e_production_based=CO2e_production_based,
+            CO2e_production_based_per_t=CO2e_production_based_per_t,
+            CO2e_total=parent.CO2e_total,
+            CO2e_total_2021_estimated=parent.CO2e_total_2021_estimated,
+            area_ha=area_ha,
+            area_ha_change=area_ha_change,
+            change_CO2e_pct=parent.change_CO2e_pct,
+            change_CO2e_t=parent.change_CO2e_t,
+            cost_climate_saved=parent.cost_climate_saved,
+            demand_change=demand_change,
+        )
+
+    @classmethod
+    def calc_soil(
+        cls, inputs: Inputs, what: str, a18: A18, area_ha: float
+    ) -> "CO2eChangeSoil":
+        CO2e_combustion_based = 0
+
+        demand_change = inputs.ass("Ass_A_P_soil_N_application_2030_change")
+        CO2e_production_based_per_t = getattr(a18, what).CO2e_production_based_per_t * (
+            1 + demand_change
+        )
+
+        CO2e_production_based = area_ha * CO2e_production_based_per_t
+
+        area_ha_change = -(getattr(a18, what).area_ha - area_ha)
+
+        parent = CO2eChange(
+            inputs=inputs,
+            what=what,
+            a18=a18,
+            CO2e_combustion_based=CO2e_combustion_based,
+            CO2e_production_based=CO2e_production_based,
+        )
+
+        return cls(
+            inputs=inputs,
+            what=what,
+            a18=a18,
+            CO2e_combustion_based=CO2e_combustion_based,
+            CO2e_production_based=CO2e_production_based,
+            CO2e_production_based_per_t=CO2e_production_based_per_t,
+            CO2e_total=parent.CO2e_total,
+            CO2e_total_2021_estimated=parent.CO2e_total_2021_estimated,
+            area_ha=area_ha,
+            area_ha_change=area_ha_change,
+            change_CO2e_pct=parent.change_CO2e_pct,
+            change_CO2e_t=parent.change_CO2e_t,
+            cost_climate_saved=parent.cost_climate_saved,
+            demand_change=demand_change,
+        )
